LayoutTests/media/media-source/media-source-real-track-enabled.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-track-enabled</title>
    <script src="mp4-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    // Real-MP4 equivalent of media-source-track-enabled.html
    // Tests that video tracks are correctly created and activeSourceBuffers populated.

    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        source = new MediaSource();
        run('video.src = URL.createObjectURL(source)');
        await waitFor(source, 'sourceopen');

        run(`sourceBuffer = source.addSourceBuffer('${MP4.VIDEO_TYPE}')`);

        var initData = MP4.initSegment({
            timescale: 1000,
            tracks: [{ id: 1, type: 'video', timescale: 1000 }]
        });
        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        testExpected('sourceBuffer.videoTracks.length', 1);
        testExpected('sourceBuffer.videoTracks[0].selected', true);
        testExpected('source.activeSourceBuffers.length', 1);

        endTest();
    }
    </script>
</head>
<body onload="runTest()">
    <video></video>
</body>
</html>
