LayoutTests/media/media-source/media-source-real-track-enabled-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer('video/mp4; codecs="avc1.4d401e"'))
EVENT(updateend)
EXPECTED (sourceBuffer.videoTracks.length == '1') OK
EXPECTED (sourceBuffer.videoTracks[0].selected == 'true') OK
EXPECTED (source.activeSourceBuffers.length == '1') OK
END OF TEST
